Moved parser interfacing to own file.
