+++ run.py
@@ -20,8 +20,8 @@ from __future__ import print_function, unicode_literals, division
 from idd3 import Relation, Engine, rules, transform
 import nltk
 from sys import argv
-from subprocess import call
 from collections import defaultdict
+from parsers import StanfordUnivDepParser
 
 import logging
 logging.basicConfig(level=logging.INFO)
@@ -47,46 +47,11 @@ except ImportError:
 
 # Stanford parser
 
-# Change this variable to the path on your system
+# Change these variable to the path on your system
 corenlp_path = os.path.expanduser('~') + \
     "/Develop/stanford_tools/corenlp"
-stanford_run_cmd = ['java', '-mx1024m', '-cp', corenlp_path + '/*:',
-                    'edu.stanford.nlp.pipeline.StanfordCoreNLP',
-                    '-annotators', 'tokenize,ssplit,pos,depparse',
-                    '-depparse.model', './data/nndep.model.txt.gz',
-                    '-outputFormat', 'conll',
-                    '-outputDirectory', '/tmp/', '-file', '-']
-
-
-def load_mapping_file(path):
-    mapping = {}
-    with open(path, 'r') as mapping_file:
-        for line in mapping_file.readlines():
-            tags = line.strip().split('\t')
-            mapping[tags[0]] = tags[1]
-
-    return mapping
-
-
-def normalize_file(infilename, pos_mapping):
-    with open('/tmp/{0}.conll'.format(infilename), 'r') as infile, \
-            open('/tmp/{0}.norm.conll'.format(infilename), 'w') as outfile:
-        for line in infile.readlines():
-            if not line or line.isspace():
-                outfile.write(line)
-            else:
-                entries = [entry.strip() for entry in line.split('\t')]
-                outfile.write('\t'.join([entries[0],
-                                         entries[1],
-                                         entries[2],
-                                         pos_mapping[entries[3]],
-                                         entries[3],
-                                         entries[4],
-                                         entries[5],
-                                         entries[6],
-                                         '_',
-                                         '_',
-                                         ]) + '\n')
+model_path = 'data/nndep.model.txt.gz'
+pos_mapping_file = 'data/ENGLISH-fine-to-universal.full.map'
 
 
 def get_sentence(graph):
@@ -134,8 +99,6 @@ def main():
         print('Usage: python', argv[0], '<input file>')
         return
 
-    pos_mapping = load_mapping_file('data/ENGLISH-fine-to-universal.full.map')
-
     if argv[1].endswith('.conll'):
         graphs = nltk.parse.dependencygraph.DependencyGraph.load(argv[1])
     else:
@@ -144,12 +107,9 @@ def main():
 
         # graphs = parser.tagged_parse_sents(tagged_sents)
 
-        stanford_run_cmd[-1] = argv[1]
-        call(stanford_run_cmd)
-        normalize_file(os.path.basename(argv[1]), pos_mapping)
-
-        graphs = nltk.parse.dependencygraph.DependencyGraph.load(
-            '/tmp/{0}.norm.conll'.format(os.path.basename(argv[1])))
+        parser = StanfordUnivDepParser(corenlp_path, model_path,
+                                       pos_mapping_file)
+        graphs = parser.parse_raw_file(argv[1])
 
     stats = process_graphs(graphs)
     print_stats(stats)
